interview 관련 dto를 하나로 합쳐라
interview 관련 dto를 하나로 합칩니다
