+++ app/src/main/java/com/example/bookclub/repository/interview/InterviewRepositoryCustom.java
@@ -1,11 +1,11 @@
 package com.example.bookclub.repository.interview;
 
-import com.example.bookclub.dto.InterviewResultDto;
+import com.example.bookclub.dto.InterviewDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface InterviewRepositoryCustom {
-	Page<InterviewResultDto> findAllContainsTileOrContent(String search, Pageable pageable);
+	Page<InterviewDto.InterviewResultDto> findAllContainsTileOrContent(String search, Pageable pageable);
 
-	Page<InterviewResultDto> findAll(Pageable pageable);
+	Page<InterviewDto.InterviewResultDto> findAll(Pageable pageable);
 }
